+++ mjenrungrot.github.io/src/index.html
@@ -6,7 +6,6 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-    <script src="./main.bundle.js"></script>
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
